backend/requirements.txt
Flask==2.3.3
pymongo==4.5.0
motor==3.3.1
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
pydantic==2.3.0
requests==2.31.0
aiohttp==3.7.4
python-dotenv==1.0.0
flask-cors==4.0.0
gunicorn==21.2.0
flask-jwt-extended==4.5.3
openai==1.3.0
google-generativeai==0.3.1
ollama==0.1.6 
